requirements.txt
Pillow==10.4.0
pyzbar==0.1.9
requests==2.25.1
smbus==1.1.post2
spidev==3.5